move pick_first tests to separate file
